+++ jest/base.config.js
@@ -5,7 +5,11 @@ module.exports = {
   coverageReporters: ['lcov'],
   modulePathIgnorePatterns: ['<rootDir>/scripts/', '<rootDir>/.*/__mocks__'],
   rootDir: path.resolve(__dirname, '..'),
-  setupFilesAfterEnv: ['<rootDir>/jest/setup/index.js'],
+  setupFilesAfterEnv: [
+    '@testing-library/jest-dom/extend-expect',
+    'jest-axe/extend-expect',
+    '<rootDir>/jest/setup/index.js'
+  ],
   transform: {
     '^.+\\.js$': '<rootDir>/jest/babel-transformer.js',
     '^.+\\.tsx?$': '<rootDir>/jest/babel-transformer.js'