+++ packages/note/src/index.ts
@@ -1,0 +1,2 @@
+export { default } from './react'
+export { default as css } from './css'