+++ packages/textarea/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [7.2.6](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-textarea@7.2.5...@pluralsight/ps-design-system-textarea@7.2.6) (2020-11-05)
+
+**Note:** Version bump only for package @pluralsight/ps-design-system-textarea
+
+
+
+
+
+## [7.2.5](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-textarea@7.2.4...@pluralsight/ps-design-system-textarea@7.2.5) (2020-11-05)
+
+**Note:** Version bump only for package @pluralsight/ps-design-system-textarea
+
+
+
+
+
 ## [7.2.4](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-textarea@7.2.3...@pluralsight/ps-design-system-textarea@7.2.4) (2020-11-02)
 
 **Note:** Version bump only for package @pluralsight/ps-design-system-textarea